app/requirements.txt
flask==2.0.1
torch==2.0.0
torchvision==0.15.0
Pillow==9.4.0
werkzeug<2.1